fix(rules): share mutation receiver alias tracking
