+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/react-builtins/no-did-mount-set-state.ts
@@ -1,4 +1,5 @@
 import { defineRule } from "../../utils/define-rule.js";
+import { collectMutationReceiverKinds } from "../../utils/collect-mutation-receiver-kinds.js";
 import type { EsTreeNode } from "../../utils/es-tree-node.js";
 import type { EsTreeNodeOfType } from "../../utils/es-tree-node-of-type.js";
 import { findEnclosingClass } from "../../utils/find-enclosing-class.js";
@@ -305,35 +306,6 @@ const objectExpressionMayDefineField = (node: EsTreeNode, fieldName: string): bo
     const propertyName = getStaticPropertyKeyName(property, { allowComputedString: true });
     return propertyName === null || propertyName === fieldName;
   });
-};
-
-const collectMutationReceiverKinds = (
-  classNode: EsTreeNode,
-): ReadonlyMap<string, "object" | "reflect"> => {
-  const receiverKinds = new Map<string, "object" | "reflect">([
-    ["Object", "object"],
-    ["Reflect", "reflect"],
-  ]);
-  let didAddReceiver = true;
-  while (didAddReceiver) {
-    didAddReceiver = false;
-    walkAst(classNode, (node) => {
-      if (
-        !isNodeOfType(node, "VariableDeclarator") ||
-        !isNodeOfType(node.id, "Identifier") ||
-        !node.init
-      ) {
-        return;
-      }
-      const initializer = stripParenExpression(node.init);
-      if (!isNodeOfType(initializer, "Identifier")) return;
-      const receiverKind = receiverKinds.get(initializer.name);
-      if (!receiverKind || receiverKinds.has(node.id.name)) return;
-      receiverKinds.set(node.id.name, receiverKind);
-      didAddReceiver = true;
-    });
-  }
-  return receiverKinds;
 };
 
 const callMayWriteThisField = (
